Math.ceil 메서드 인수 값 형변환 타입 변경

+++ programmers/난이도별/level02.예상_대진표/SSSOy.java
@@ -8,8 +8,8 @@ class Solution {
         }
         
         while (!(b - a == 1 && b % 2 == 0)) {
-            a = (int) Math.ceil((float) a / 2);
-            b = (int) Math.ceil((float) b / 2);
+            a = (int) Math.ceil((double) a / 2);
+            b = (int) Math.ceil((double) b / 2);
             answer++;
         }
